fixed missing ctype.h include and put back casting to APP_DEBUG in config.c

+++ public/step/samples/device_request/device_request.c
@@ -97,12 +97,12 @@ static idigi_callback_status_t app_process_device_request(idigi_data_service_msg
         if (client_device_request->target != NULL)
         {
             /* target is only available on 1st chunk of data */
-            APP_DEBUG("Device request data: received data = \"%.*s\" for target = \"%s\"\n", server_data->length_in_bytes,
+            APP_DEBUG("Device request data: received data = \"%.*s\" for target = \"%s\"\n", (int)server_data->length_in_bytes,
                     device_request_data, client_device_request->target);
         }
         else
         {
-            APP_DEBUG("Device request data: received data = \"%.*s\" for unknown target\n", server_data->length_in_bytes,
+            APP_DEBUG("Device request data: received data = \"%.*s\" for unknown target\n", (int)server_data->length_in_bytes,
                     device_request_data);
 
         }
@@ -137,7 +137,7 @@ static idigi_callback_status_t app_process_device_response(idigi_data_service_ms
         /* get number of bytes written to the client data buffer */
         size_t const bytes = (client_device_request->length_in_bytes < client_data->length_in_bytes) ? client_device_request->length_in_bytes : client_data->length_in_bytes;
 
-        APP_DEBUG("Device response data: send response data = %.*s\n", bytes, client_device_request->response_data);
+        APP_DEBUG("Device response data: send response data = %.*s\n", (int)bytes, client_device_request->response_data);
 
         /* let's copy the response data to service_response buffer */
         memcpy(client_data->data, client_device_request->response_data, bytes);
